+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## master
+
+### Added
+
+- `Raxx.set_secure_browser_headers/1` Adds a collection of useful headers when responding to a browser.
+
 ## [0.14.13](https://github.com/CrowdHailer/raxx/tree/0.14.13) - 2018-03-19
 
 ### Added